feat(core): adds title to document panel
