+++ packages/sanity/src/desk/panes/document/documentPanel/documentViews/FormView.tsx
@@ -1,10 +1,11 @@
 /* eslint-disable no-nested-ternary */
 
-import {Box, Container, Flex, Spinner, Text, focusFirstDescendant} from '@sanity/ui'
-import React, {forwardRef, useEffect, useMemo, useRef, type Ref, useCallback, useState} from 'react'
+import {Box, Container, Flex, Heading, Spinner, Text, focusFirstDescendant} from '@sanity/ui'
+import React, {forwardRef, useEffect, useMemo, useCallback, useState} from 'react'
 import {tap} from 'rxjs/operators'
 import {useDocumentPane} from '../../useDocumentPane'
 import {Delay} from '../../../../components/Delay'
+import {useDocumentTitle} from '../../useDocumentTitle'
 import {useConditionalToast} from './useConditionalToast'
 import {
   DocumentMutationEvent,
@@ -49,6 +50,7 @@ export const FormView = forwardRef<HTMLDivElement, FormViewProps>(function FormV
   } = useDocumentPane()
   const documentStore = useDocumentStore()
   const presence = useDocumentPresence(documentId)
+  const {title} = useDocumentTitle()
 
   // The `patchChannel` is an INTERNAL publish/subscribe channel that we use to notify form-builder
   // nodes about both remote and local patches.
@@ -149,6 +151,11 @@ export const FormView = forwardRef<HTMLDivElement, FormViewProps>(function FormV
     >
       <PresenceOverlay margins={margins}>
         <Box as="form" onSubmit={preventDefault} ref={setRef}>
+          <Box marginBottom={4}>
+            <Heading muted={!title} size={5}>
+              {title ?? 'Untitled'}
+            </Heading>
+          </Box>
           {ready ? (
             formState === null ? (
               <Box padding={2}>
