+++ packages/sanity/src/desk/panes/document/documentPanel/header/DocumentHeaderTitle.tsx
@@ -1,32 +1,12 @@
 import React, {ReactElement} from 'react'
-import {useDocumentPane} from '../../useDocumentPane'
-import {unstable_useValuePreview as useValuePreview} from 'sanity'
+import {useDocumentTitle} from '../../useDocumentTitle'
 
 export function DocumentHeaderTitle(): ReactElement {
-  const {connectionState, schemaType, title, value: documentValue} = useDocumentPane()
-  const subscribed = Boolean(documentValue) && connectionState === 'connected'
-
-  const {error, value} = useValuePreview({
-    enabled: subscribed,
-    schemaType,
-    value: documentValue,
-  })
-
-  if (connectionState !== 'connected') {
-    return <></>
-  }
-
-  if (title) {
-    return <>{title}</>
-  }
-
-  if (!documentValue) {
-    return <>New {schemaType?.title || schemaType?.name}</>
-  }
+  const {error, title} = useDocumentTitle()
 
   if (error) {
-    return <>Error: {error.message}</>
+    return <>{error}</>
   }
 
-  return <>{value?.title || <span style={{color: 'var(--card-muted-fg-color)'}}>Untitled</span>}</>
+  return <>{title || <span style={{color: 'var(--card-muted-fg-color)'}}>Untitled</span>}</>
 }
