+++ packages/sanity/src/desk/panes/document/index.ts
@@ -1,2 +1,3 @@
 export {DocumentPane as default} from './DocumentPane'
 export * from './DocumentPaneProvider'
+export * from './useDocumentTitle'
